requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
reportlab==4.0.7
python-docx==1.1.0
Pillow==10.1.0
email-validator==2.1.0
bcrypt==4.1.2

# Security & Rate Limiting
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1

# Production Server
gunicorn==21.2.0
psycopg2-binary==2.9.9

# Caching
Flask-Caching==2.1.0
redis==5.0.1

# Monitoring & Logging
sentry-sdk[flask]==1.38.0
prometheus-flask-exporter==0.23.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
flask-shell-ipython==1.4.0
